% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{olympics_participation}
\alias{olympics_participation}
\title{Olympic Participation Data}
\format{
\subsection{'olympics_participation'}{

A data frame with A data frame with 271116 rows and 15 variables
\describe{
\item{id}{Unique identifier for each athlete}
\item{name}{Athlete's name}
\item{sex}{Athlete's gender}
\item{age}{Athlete's age at the time of the Olympics}
\item{height}{Athlete's height in centimeters}
\item{weight}{Athlete's weight in kilograms}
\item{team}{The country or team the athlete represented}
\item{noc}{National Olympic Committee region}
\item{games}{Name of the Olympic games}
\item{year}{Year of the Olympics}
\item{season}{Season of the Olympics}
\item{city}{Host city}
\item{sport}{Sport category}
\item{event}{Specific event}
\item{medal}{Medal won (Gold, Silver, Bronze, or NA)}
}
}
}
\source{
\url{https://github.com/rfordatascience/tidytuesday/blob/master/data/2024/2024-08-06/readme.md}
}
\usage{
olympics_participation
}
\description{
This dataset contains historical participation data from the Summer and Winter Olympics,
focusing on male and female athletes from 1900 to 2016.
}
\keyword{datasets}
